requirements.txt
streamlit
pandas
plotly
python-dotenv
pydantic
fastapi
uvicorn
requests
google-genai
python-dotenv